reuse relationship panels right under detail view instead of recreating them on each refresh. reuse labels in when refreshing detail view. only start deleting (ask for deletion) if at least one item has been selected. labeled delete button with 'delete selected' and import button with 'import csv'. removed export buttons since export is not integrated yet.

+++ branches/honeycrm_r134_gwt2.1m2/src/honeycrm/client/view/ListView.java
@@ -60,12 +60,14 @@ public class ListView extends AbstractView {
 	}
 
 	private Button getDeleteButton() {
-		final Button btn = new Button("Delete");
+		final Button btn = new Button("Delete selected");
 		btn.addClickHandler(new ClickHandler() {
 			@Override
 			public void onClick(ClickEvent event) {
-				if (Window.confirm("Do you want to delete the selected items?")) {
-					deleteSelected(getDeletedIds());
+				if (getDeletedIds().size() > 0) {
+					if (Window.confirm("Do you want to delete the selected items?")) {
+						deleteSelected(getDeletedIds());
+					}
 				}
 			}
